mass refactor on codebase structure
