--- src/app/index.tsx
@@ -1,6 +1,0 @@
-export { Header } from './header/Header';
-export { Content } from './content/Content';
-export { SidebarContainer } from './sidebar/containers/SidebarContainer';
-export { SchedulerContainer } from './scheduler/SchedulerContainer';
-export { SchedulerSidebar } from './scheduler/components/SchedulerSidebar';
-export { Feedback } from './feedback/Feedback';
