--- src/app/registration/index.tsx
@@ -1,1 +1,0 @@
-export { RegistrationContainer as Registration } from './containers/RegistrationContainer';
