+++ src/common/sidebar/components/CoursesList.tsx
