+++ src/pages/calendar/Content.stories.tsx
@@ -2,7 +2,7 @@ import { Story, Meta } from '@storybook/react/types-6-0';
 import React from 'react';
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 
-import { Content, ContentProps } from './Content';
+import { Content, ContentProps } from './containers/Content';
 
 export default {
   title: 'Content',
