+++ src/pages/home/containers/Landing.tsx
@@ -2,7 +2,7 @@ import { Image } from '@chakra-ui/image';
 import { Container, Flex, Grid, GridItem, Heading, Text } from '@chakra-ui/layout';
 import { useEffect, useState } from 'react';
 
-export default function Landing() {
+export function Landing() {
   const [handIndex, setHandIndex] = useState(0);
   const hands = ['👋', '👋🏻', '👋🏼', '👋🏽', '👋🏾', '👋🏿'];
