+++ src/pages/registration/components/RegistrationMinimized.tsx
