+++ src/pages/registration/containers/CourseContainer.tsx
