+++ src/pages/registration/containers/RegistrationContainer.tsx
@@ -2,10 +2,10 @@ import { Box, Divider, Flex, Heading } from '@chakra-ui/layout';
 import { Helmet } from 'react-helmet';
 import { useParams } from 'react-router';
 
+import { Feedback } from '../../../common/feedback/Feedback';
+import { Header } from '../../../common/header/Header';
 import { useSavedCourses } from '../../../shared/hooks/useSavedCourses';
-import { Feedback } from '../../feedback/Feedback';
-import { Header } from '../../header/Header';
-import { getReadableTerm } from '../../shared/utils/terms';
+import { getReadableTerm } from '../../../shared/utils/terms';
 import { RegistrationHeading } from '../components/RegistrationHeading';
 
 import { CourseContainer } from './CourseContainer';
