+++ src/pages/scheduler/components/CourseCard.tsx
