+++ src/shared/utils/index.ts
@@ -1,0 +1,3 @@
+export * from './dates';
+export * from './terms';
+export * from './courses';
