+++ src/shared/utils/terms.ts
